[wdg] Move Viewport to the Widgets subfolder.
