+++ code/Editors/GamePlayer/GamePlayer.cpp
@@ -21,7 +21,7 @@
 
 #include "Widgets/Button.h"
 #include "Widgets/Layout.h"
-#include "Widgets/Viewport.h"
+#include "Widgets/Widgets/Viewport.h"
 
 namespace Editors
 {
