+++ code/Editors/LevelEditor/Widgets/LevelEditorViewportWidget.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "Widgets/Viewport.h"
+#include "Widgets/Widgets/Viewport.h"
 
 namespace Rendering
 {
