+++ code/Editors/MaterialEditor/MaterialEditor.cpp
@@ -53,8 +53,8 @@
 #include "Widgets/Tab.h"
 #include "Widgets/TabContainer.h"
 #include "Widgets/Text.h"
-#include "Widgets/Viewport.h"
 #include "Widgets/Widgets/TableView.h"
+#include "Widgets/Widgets/Viewport.h"
 
 #include <filesystem>
