+++ code/Editors/TextureEditor/TextureEditor.cpp
@@ -45,9 +45,9 @@
 #include "Widgets/Models/SelectionRow.h"
 #include "Widgets/SplitHorizontal.h"
 #include "Widgets/SplitVertical.h"
-#include "Widgets/Viewport.h"
 #include "Widgets/WidgetMgr.h"
 #include "Widgets/Widgets/TableView.h"
+#include "Widgets/Widgets/Viewport.h"
 
 #include <chrono>
