+++ code/Widgets/Widgets.vcxproj
@@ -50,7 +50,7 @@
     <ClInclude Include="Tools\Color.h" />
     <ClInclude Include="Tools\DefaultColors.h" />
     <ClInclude Include="Tools\Rect.h" />
-    <ClInclude Include="Viewport.h" />
+    <ClInclude Include="Widgets\Viewport.h" />
     <ClInclude Include="Widget.h" />
     <ClInclude Include="WidgetMgr.h" />
     <ClInclude Include="Widgets\AbstractClickable.h" />
@@ -95,7 +95,7 @@
     <ClCompile Include="Tools\Color.cpp" />
     <ClCompile Include="Tools\DefaultColors.cpp" />
     <ClCompile Include="Tools\Rect.cpp" />
-    <ClCompile Include="Viewport.cpp" />
+    <ClCompile Include="Widgets\Viewport.cpp" />
     <ClCompile Include="Widget.cpp" />
     <ClCompile Include="WidgetMgr.cpp" />
     <ClCompile Include="Widgets\AbstractClickable.cpp" />
